Support https, too
More complicated than expected, even if the actual code added are not many
lines.

Figuring out the exact Poco code needed to initialise Poco's OpenSSL bits the
right way was confusing. Which is odd, one would thing that it would be a most
common use case for OpenSSL, to be able to download documents over https, and
that there would be some trivial way to tell Poco to do set things up for
that.