Update patch set 2
Patch Set 2:

This bug is the case itself :)

If you do a dry-run on 'nextmessage' frame going in prisoner -> master direction, you would notice that it would get stuck in the original if/else block because 'nextmessage' is a single-line message also. Hence, it is necessary for us to move this block above so that such 'nextmessage' frames are not skipped.

I think it was always wrongly placed there, but we were never able to notice it because only few documents hit the 200000 limit (char buffer[200000] in SocketProcessor() function ).

Patch-set: 2
Reviewer: Gerrit User 1000897 <1000897@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0