Update patch set 2
Patch Set 2:

(1 comment)

> I think it was always wrongly placed there, but we were never able
 > to notice it because only few documents hit the 200000 limit (char
 > buffer[200000] in SocketProcessor() function ).

The assumption was that the only multi-line message was nextmessage. That is, everything else evaluated to single-line, so the sequence was reasonable (that's my understanding).

At any rate, I think we need to review these assumptions. For example the code now expects 3 cases: nextmessage, single-line, and a 3rd undocumented case.

Please add the const (see note) and I'll submit this patch, then we need to review the protocol(.txt) and the code and make sure we have a solid situation. Thanks.

Patch-set: 2