Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

I've added a few comments, please see inline.

But the biggest issue I have is with the design. We should avoid unnecessary complexity. When the user opens a password-protected document (without providing a password, or with wrong password) it will fail. Loading will fail. End of story.

The UI should report the failure and prompt for password. Another load request should then be sent to the server, with the password, to try loading again.

There is no need to have a password command, or remembering the jail url, or having a new callback (modify load callback to take the optional password) etc. All we need is for 'load' to have an optional password that is passed to the right API when provided.

If 'load' fails, then everything fails and the user can try again. It's a fatal error to load a document and without a document there is no need to maintain anything at all (certainly no child). I'd even close the socket and cleanup everything. The UI should handle this case in a friendly fashion.
Fail, return, prompt the user, and let them try again.

Let's discuss if there are any design questions. Thanks!

Patch-set: 3
Label: Code-Review=-1