Update patch set 3
Patch Set 3:

Ashod: Hmm, the design concern is actually a good point; unfortunately on the LibreOffice side, it asks for the password interactively, so the LOK API was (naturally) extended so that we can interactively answer.

Having said that, on the loolwsd level, it can be done as you suggest - return from load with "no password provided" error, ask the user for the password, and re-try the load completely (now with the password remembered).  That is safer (the load is not blocked), but OTOH more expensive (2x starting the load).

Maybe we can even change the LOK API this way; but not sure what is best :-(

Michael S.: Any view on this, please?

Patch-set: 3
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0