Update patch set 4
Patch Set 4:

> (7 comments)
 > What stopped me from submitting is strdup :)
 > I'll guess it was muscle memory :)
 > 
 > Seriously, why not use std::string instead of the naked char*?

My only concern was documents having password as empty string, "", and std::string's default initialization to "". 

Also, LOK's setDocumentPassword function treats an empty string as a password attempt, and would get triggered again unless I pass a nullptr to it to abort the load operation, but of course that could easily be dealt with.

If there are no empty strings passwords for documents, then yeah, we can change it to std::string, assuming "" to be no password provided

Patch-set: 4