Update patch set 4
Patch Set 4:

> > (7 comments)
 > > What stopped me from submitting is strdup :)
 > > I'll guess it was muscle memory :)
 > >
 > > Seriously, why not use std::string instead of the naked char*?
 > 
 > My only concern was documents having password as empty string, "",
 > and std::string's default initialization to "".
 > 

Reasonable. It's probably better to have a bool to track if we have a password (i.e. the user has provided a password) rather than track a pointer (and free). Btw, I don't think I noticed a call to free() anywhere.

If we clean this part, I'm good to submit and your other comments are reasonable and can be dealt with later.

Patch-set: 4