Update patch set 7
Patch Set 7:

The main reason to wait for load acknowledgment and issue commands only after 'statusindicatorfinish' is received is also how we currently handle the docalreadyloaded, nodocloaded cases on the server side. Currently we have a very lame condition on the server side checking for docURL. If docURL != "" for the server, it means doc is loaded which is clearly not true since the load operation can fail.

Hence, IMO, it is best to wait for the result of such an important command like 'load' to finish before we can issue commands from the front-end.

Patch-set: 7
Label: Verified=0