loolwsd: Websocket subscription of AdminModel updates
After opening the websocket, it is added to the list of
subscribers in AdminModel. Any further updates to AdminModel
notifies the list of subscribers. If there are no admin panel
sessions, then all the notification will not be sent over the
websocket, but still it would update the AdminModel which can be
queried later on by admin panel sessions to fill up the admin UI.

Change-Id: I2fbfdb4dbf600283b154393a7f06c6b3c0163753