Update patch set 1
Patch Set 1:

Pranav: Thanks so much for this!  I just wonder about the directory layout - the 'dist/' is supposed to be the place that you deploy at the end, do we really need everything from the bootstrap/ dir there?  Shouldn't it be that it is in the src, and just things like bootstrap.min.css are copied to dist/ during make?

OTOH debug/document/dashboard.html looks like something that should be in the dist/ ?

Patch-set: 1