Update patch set 3
Patch Set 3:

> (1 comment)

These certificate files are supposed to be generated by the sysadmin. For the convenience of testing and debugging, I have included the sample cert and key files in the code base now. Ideally, this path would be picked up from the configuration file that would be read when loolwsd is started, but configuration file is a WIP, I guess.

Patch-set: 3
Reviewer: Gerrit User 1000897 <1000897@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0