Update patch set 3
Patch Set 3:

> But having such "sample" certificates in the source code *will*
 > lead to some clueless person eventually using them for their
 > small-business lool setup, and then wondering how comes they got
 > pwned so quickly.

:)

Yeah, as soon as we have configuration file, I will remove them straight away. Its just that this SSL thing would be blocked/would be difficult to test if I do not keep them there for the time being. Just wanted to make it 'build and run' kind of setup for now :)

Patch-set: 3