loleaflet: Clean-up toolbar; fix en/dis-able button states
Use an object to keep track of enabled/disabled state of buttons.
This also prevents us looping all the format buttons everytime a
commandstatechanged is received, so better algorithmic
complexity.

For sessions in view-only mode, keep storing the state of these
buttons in this object, and use the stored states to
enable/disable buttons when edit lock is taken.

While enabling/disabling the toolbar buttons, also check in
toolbar-more.

Change-Id: I6555fde6669e1a8ba349c3213280a627a05d80b7