Update patch set 2
Patch Set 2: Code-Review-1

(8 comments)

Some comments inline.

I think if the key sending functions are moved to helpers.hpp, we can commit the invalidatetile test.

The other "monkey test" code, I'm not sure what it's achieving. Perhaps a comment to explain is best. Also, it should go into its own test function and committed separately from the invalidatetile (and it probably doesn't belong in this file anyway (should be in httpwstest.cpp).

Regarding the empty docs, the typical way of getting an empty do is to load one of the existing, selectall, delete. But I guess it doesn't hurt to have empty ones.

P.S. If code is for review-only, you can signal that it's not for merging by giving it a Code-Review score of -2, then add reviewers and mention the purpose. Less confusion.

Patch-set: 2
Label: Code-Review=-1