bccu#1909 - loolwsd command line options vs. config file and package upgrades
Default config entries and values are now set before loading
the config file from disk. This ensures that an upgrade
to the server with new config entries will not break the
server when the XML is not upgrade, rather, the server
will fallback to the defaults when it fails to find the
entry in the XML.

This, of course, requires adding new entries to the
default entries every time config entries are changed.

Change-Id: Iac849dc41fc7410a3c9093b0d6e91d96182b86fd