Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

The idea to merge the two threads (callback handler and tile rendering) is reasonable, but we should discuss the approach.

The starvation you noticed happens because you are using the tile-queue, which is designed for... well, yes, tiles! :)

It reorders the messages such that it renders tiles in large blocks first. So with heavy typing the callback messages always end up getting overtaken by tiles, which the queue explicitly favors, thereby starving callback events.

So this is one problem with the current implementation. I think the tiles should have its own queue, and callbacks a separate one. The thread that pumps them would process one callback, then one tile-rendering.

Another issue with combining these threads is that there will be reduction in performance. This is because before and after rendering a tile there is significant work being done that does *not* execute with the Solar lock. This code (mostly encoding to PNG) is CPU intensive and can run in parallel to other code (even to other rendering calls).

Finally, I have submitted my patch (342125b8ea) to fix the missing tiles issue. Should be easier to read than describe it here (it's only +27, -6). It comes with detailed commit log too. So this approach is mostly an optimization to reduce the thread-count, but, as mentioned, we need multiple threads because we have opportunity to parallelize, and merging them will reduce performance (unless benchmark proves otherwise, of course).

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1