Update patch set 1
Patch Set 1:

Wrt. the starvation - actually that was due to the logging; we have so much debugging stuff that it was blocking considerably when writing to the console, making the queue grow.  Without that, all works fine even with the unmodified TileQueue :-)

But I worked on the TileQueue too - it seems unnecessary to do all the priority shuffling when inserting, when the prioritization changes with every view switch, and instead do it only when the tile is requested - which seems to work nicely.

Wrt. an own queue for tiles - that is exactly the thing that makes the synchronzation hard, as then it is necessary to synchronize between those 2 queues; which in turn kills all the advantage of 2 queues :-)

Wrt. PNG compression; if we want to thread it, then let's thread just the compression - that'll make it local, more maintainable, and we can go with as many threads there as we wish (for just the PNG compression piece of code).

Wrt. the 342125b8ea - thanks for that!  Unfortunately that again just adds code :-(

Either way, please check https://cgit.freedesktop.org/libreoffice/online/log/?h=private/kendy/tilesandcallbacks , I'd really like to get that in...  I've tested it with several people in a session, and it performed pretty well thanks to all your work I was able to build on.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0