Update patch set 2
Patch Set 2:

> I'm not sure I understand why broadcasting of tiles should be decided by a client.

If you can figure out a clever way for the server to deduce which tile requests are for previews after the client has edited a slide (and not for previews when the document is being opened, or other reasons), then sure, the logic could be in the server, too. I tend to favour pragmatic solutions that works when I try them.

> each client can decide for themselves (whether they want to get unrequested tiles, push style)?

But how would a client A know that some other client B has edited a slide that client A is showing a preview of? We would need some new message for that, or start sending invalidate messages for slides that the client is not even editing at the moment, or something.

> What if some clients set broadcast=no? Should that work fine?

That is like what is done now and should not change anything.

Patch-set: 2