Update patch set 2
Patch Set 2:

> > I'm not sure I understand why broadcasting of tiles should be
 > decided by a client.
 > 
 > If you can figure out a clever way for the server to deduce which
 > tile requests are for previews after the client has edited a slide
 > (and not for previews when the document is being opened, or other
 > reasons), then sure, the logic could be in the server, too. I tend
 > to favour pragmatic solutions that works when I try them.

I think it's better to identify slides in a documented and friendly way (I'm not a fan of id=?? as a way to recognize slides) and then the server can always push these to all views.

But a simple solution would be for each client to automatically issue a slide tile request when any tile is invalidated in a given slide.

Why can't we do that?

Patch-set: 2